fsm: GetState and SetState via applying build to a Stater type
This commit adds a new Stater interface which is a type that is included during
a Builder.Build.  Build will use Stater's SetState to set the resulting state
after a Transition. Also, Build uses the Stater's GetState to set the initial
state for the FSM.

Also, this commit introduces Stater as an argument to all handler callback
functions so that the Stater can be acted upon or have other serialized
components of a Transition be persisted as well.